[AoC-2021] - Day 2.

+++ src/main/java/io/exercises/Day.java
@@ -1,6 +1,7 @@
 package io.exercises;
 
-public interface Day {
-    int part1(String input);
-    int part2(String input);
+public interface Day
+{
+    long part1(String input);
+    long part2(String input);
 }
